Add diskutilities module
